Add write_files.py
